requirements.txt
# Основные зависимости для Telegram-бота

# Aiogram 3.x - фреймворк для Telegram ботов
aiogram>=3.0.0

# Асинхронные HTTP запросы
aiohttp>=3.9.0

# Парсинг HTML
beautifulsoup4>=4.12.0

# Парсер для BeautifulSoup
lxml>=5.0.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# Работа с часовыми поясами
pytz>=2024.1
